fixed bug in file_counter
If there were files in the directory already modified then file counter would return number of modified files. zero_padder() then used this number for the length comparison instead of total number of files in directory. if a situation with a directory of more than 99999+named files to rename but only  a small number of already named files  occurred then this would result in an incorrect zfill()